Prepare HiSSE marginal reconstruction for plotting

A function to process the output from hisse::MarginRecon. First we get model averaged rates for tips and nodes of the phylogeny. Then convert the tree from phylo format into a treedata class and assign the data.frame of mode-averaged rates to the @data slot. We output the node and tip rates tables and the reformated tree object for downstream plotting with ggtree and ggplot.

h_process_recon(hisse_recon)

Arguments

hisse_recon
An object produced by hisse::MarginRecon, or a list of such objects over which we can model-average using hisse::GetModelAveRates

Value

A list with three components:

tree_data the phylogeny with associated data object with reconstructed ancestral, model-averaged node and tip states and rates

tip_rates/node_rates a tibble of reconstructed ancestral, model-averaged tip/node states and rates.

Examples

data("diatoms")
processed_hisse <- h_process_recon(hisse_recon=diatoms$cid4_recon)
processed_hisse$tip_rates

#> # A tibble: 1,132 x 7
#>    id                  state turnover net.div speciation extinct.frac extinction
#>    <chr>               <dbl>    <dbl>   <dbl>      <dbl>        <dbl>      <dbl>
#>  1 acanthoceras_sp_hk…     0   0.0717  0.0167     0.0442        0.290     0.0275
#>  2 urosolenia_eriensi…     0   0.0717  0.0167     0.0442        0.290     0.0275
#>  3 dactyliosolen_blav…     0   0.0702  0.0167     0.0435        0.282     0.0268
#>  4 hemiaulus_sinensis…     0   0.0702  0.0167     0.0435        0.282     0.0268
#>  5 chaetoceros_brevis      0   0.0802  0.0165     0.0483        0.321     0.0318
#>  6 chaetoceros_debili…     0   0.0802  0.0165     0.0483        0.321     0.0318
#>  7 chaetoceros_calcit…     0   0.0839  0.0163     0.0501        0.356     0.0338
#>  8 chaetoceros_curvis…     0   0.373   0.0180     0.195         0.647     0.177 
#>  9 chaetoceros_muelle…     0   0.373   0.0180     0.195         0.647     0.177 
#> 10 chaetoceros_debili…     0   0.577   0.0203     0.299         0.732     0.278 
#> # … with 1,122 more rows

processed_hisse$node_rates

#> # A tibble: 1,131 x 7
#>    id        state turnover net.div speciation extinct.frac extinction
#>    <chr>     <dbl>    <dbl>   <dbl>      <dbl>        <dbl>      <dbl>
#>  1 1133  0.00350       1.11  0.0239      0.569        0.876      0.545
#>  2 1134  0.00185       1.29  0.0255      0.659        0.942      0.634
#>  3 1135  0.00746       1.38  0.0264      0.703        0.962      0.677
#>  4 1136  0.0154        1.40  0.0267      0.711        0.963      0.684
#>  5 1137  0.542         1.44  0.0276      0.736        0.963      0.708
#>  6 1138  0.544         1.45  0.0278      0.740        0.963      0.712
#>  7 1139  0.404         1.47  0.0282      0.750        0.962      0.722
#>  8 1140  0.379         1.50  0.0289      0.767        0.959      0.738
#>  9 1141  0.0202        1.57  0.0301      0.798        0.960      0.768
#> 10 1142  0.0000388     1.61  0.0309      0.820        0.962      0.789
#> # … with 1,121 more rows

processed_hisse$tree_data

#> 'treedata' S4 object'.
#> 
#> ...@ phylo: 
#> Phylogenetic tree with 1132 tips and 1131 internal nodes.
#> 
#> Tip labels:
#>   acanthoceras_sp_hk250, urosolenia_eriensis_hk054, dactyliosolen_blavyanus_hk301, hemiaulus_sinensis_hk123, chaetoceros_brevis, chaetoceros_debilis_ch48, ...
#> Node labels:
#>   4, 4, 4, 4, 8, 8, ...
#> 
#> Rooted; includes branch lengths.
#> 
#> with the following features available:
#> 	'id',	'state',	'turnover',	'net.div',	'speciation',	'extinct.frac',
#> 	'extinction'.