Extract transition and diversification rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)

This is a function that wraps m_transition_matrix and m_diversification_rates so rates can be extracted in one step

m_collect_rates(model_fit, hidden_traits, character_states)

Arguments

model_fit
A muhisse.fit object

hidden_traits
Logical indicating whether the model has hidden states

character_states
A translation for the character states in the model in the order 00, 01, 10, 11

Value

A list with two components, the transition rates formatted as a rate matrix and a data frame of diversification rates

Examples

States <- c("mp", "mb", "fp", "fb")

# muhisse model
data("diatoms")
m_collect_rates(model_fit = diatoms$muhisse, hidden_traits=TRUE, character_states=States)

#> $Transition_rates
#>              mp1          mb1         fp1          fb1         mp2         mb2
#> mp1           NA 2.061154e-09 0.000125180 0.000000e+00 0.002832762          NA
#> mb1 7.971849e-04           NA 0.000000000 1.617498e-03          NA 0.002832762
#> fp1 2.061154e-09 0.000000e+00          NA 2.061154e-09          NA          NA
#> fb1 0.000000e+00 3.834735e-03 0.001416519           NA          NA          NA
#> mp2 6.483053e-02           NA          NA           NA          NA 0.002568749
#> mb2           NA 6.483053e-02          NA           NA 0.003050672          NA
#> fp2           NA           NA 0.064830535           NA 0.010208014 0.000000000
#> fb2           NA           NA          NA 6.483053e-02 0.000000000 0.002309284
#>              fp2         fb2
#> mp1           NA          NA
#> mb1           NA          NA
#> fp1 2.832762e-03          NA
#> fb1           NA 0.002832762
#> mp2 4.316343e-03 0.000000000
#> mb2 0.000000e+00 0.011032817
#> fp2           NA 0.031237128
#> fb2 2.061154e-09          NA
#> 
#> $Diversification_rates
#> # A tibble: 8 x 6
#>   State Turnover Extinction_fraction Speciation Extinction Net_diversification
#>   <chr>    <dbl>               <dbl>      <dbl>      <dbl>               <dbl>
#> 1 mp1     0.0976               0.771     0.0551     0.0425              0.0126
#> 2 mb1     0.0977               0.771     0.0552     0.0425              0.0126
#> 3 fp1     0.204                0.771     0.115      0.0887              0.0264
#> 4 fb1     0.170                0.771     0.0961     0.0741              0.0220
#> 5 mp2     0.615                0.771     0.347      0.268               0.0795
#> 6 mb2     0.569                0.771     0.321      0.248               0.0736
#> 7 fp2     1.42                 0.771     0.802      0.618               0.184 
#> 8 fb2     0.987                0.771     0.557      0.430               0.128 
#>