Prepare MuHiSSE marginal reconstruction for plotting

A function to process the output from hisse::MarginReconMuHiSSE. First we get model averaged rates for tips and nodes of the phylogeny and probabilities from the ancestral reconstruction of character states. Then convert the tree from phylo format into a treedata class and assign the data.frame of mode-averaged rates to the @data slot. We output the node and tip rates tables and the reformated tree object for downstream plotting with ggtree and ggplot.

m_process_recon(muhisse_recon)

Arguments

muhisse_recon
An object produced by hisse::MarginReconMuHiSSE, or a list of such objects over which we can model-average using hisse::GetModelAveRates

Details

Internally, the probabilities for individual states (00, 01, 10, 11) are converted into prob_0x giving the probability for the first character being 0, and prob_x0 giving the probability of the second character being 0. This way we can plot in terms of a 'focal character', i.e., we can plot the rate on the y axis and the probability of being 0 for the 'focal character' on the x axis while color-coding by the states of both characters.

#'@return A list with three components:

tree_data the phylogeny with associated data object with reconstructed ancestral, model-averaged node and tip states and rates

tip_rates/node_rates a tibble of reconstructed ancestral, model-averaged tip/node states and rates.

Examples

data("diatoms")
processed_muhisse <- m_process_recon(muhisse_recon=diatoms$muhisse_recon)
processed_muhisse$tip_rates

#> # A tibble: 1,132 x 12
#>    id    state.00 state.01 state.10 state.11 turnover net.div speciation
#>    <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>      <dbl>
#>  1 acan…        0        0        1        0    0.228  0.0281     0.128 
#>  2 uros…        0        0        1        0    0.228  0.0281     0.128 
#>  3 dact…        1        0        0        0    0.111  0.0137     0.0622
#>  4 hemi…        1        0        0        0    0.111  0.0137     0.0622
#>  5 chae…        1        0        0        0    0.111  0.0137     0.0625
#>  6 chae…        1        0        0        0    0.111  0.0137     0.0625
#>  7 chae…        1        0        0        0    0.111  0.0137     0.0622
#>  8 chae…        1        0        0        0    0.254  0.0314     0.143 
#>  9 chae…        1        0        0        0    0.254  0.0314     0.143 
#> 10 chae…        1        0        0        0    0.371  0.0459     0.209 
#> # … with 1,122 more rows, and 4 more variables: extinct.frac <dbl>,
#> #   extinction <dbl>, prob_0x <dbl>, prob_x0 <dbl>

processed_muhisse$node_rates

#> # A tibble: 1,131 x 12
#>    id    state.00 state.01 state.10 state.11 turnover net.div speciation
#>    <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>      <dbl>
#>  1 1133  0.870    0.128     1.78e-3  3.33e-4    0.556  0.0686      0.312
#>  2 1134  0.888    0.112     2.42e-4  1.18e-5    0.623  0.0770      0.350
#>  3 1135  0.0845   0.912     1.03e-3  2.75e-3    0.594  0.0734      0.334
#>  4 1136  0.0776   0.922     1.04e-4  1.26e-4    0.592  0.0731      0.333
#>  5 1137  0.00105  0.999     2.98e-7  1.88e-5    0.589  0.0727      0.331
#>  6 1138  0.000415 1.00      7.58e-8  6.86e-6    0.589  0.0727      0.331
#>  7 1139  0.00286  0.997     6.64e-7  4.88e-5    0.587  0.0726      0.330
#>  8 1140  0.0224   0.977     3.79e-5  1.95e-4    0.586  0.0724      0.329
#>  9 1141  0.953    0.0471    4.83e-5  3.23e-6    0.635  0.0785      0.357
#> 10 1142  1.00     0.000318  1.37e-4  1.03e-8    0.640  0.0790      0.359
#> # … with 1,121 more rows, and 4 more variables: extinct.frac <dbl>,
#> #   extinction <dbl>, prob_0x <dbl>, prob_x0 <dbl>

processed_muhisse$tree_data

#> 'treedata' S4 object'.
#> 
#> ...@ phylo: 
#> Phylogenetic tree with 1132 tips and 1131 internal nodes.
#> 
#> Tip labels:
#>   acanthoceras_sp_hk250, urosolenia_eriensis_hk054, dactyliosolen_blavyanus_hk301, hemiaulus_sinensis_hk123, chaetoceros_brevis, chaetoceros_debilis_ch48, ...
#> Node labels:
#>   5, 5, 6, 6, 6, 6, ...
#> 
#> Rooted; includes branch lengths.
#> 
#> with the following features available:
#> 	'id',	'state.00',	'state.01',	'state.10',	'state.11',	'turnover',	'net.div',
#> 	'speciation',	'extinct.frac',	'extinction',	'prob_0x',	'prob_x0'.