Extract transition rates from a multistate HiSSE model (MuSSE, MuHiSSE or CID)

A helper function that formats hisse::MuHiSSE's output into a matrix so transition rates between states can be easier to visualize.

m_transition_matrix(model_fit, hidden_states = TRUE, states)

Arguments

model_fit
A muhisse fit object

hidden_states
Logical whether the model contains hidden states

states
A vector of state labels to replace hisse's 0/1 notation (see examples)

Value

The estimated transition rates formatted in a matrix

Examples

# The translation should follow the order of states in model_fit$solution.
# In this example, hisse's 00, 01, 10, and 11 states correspond to
# marine-plankton, marine-benthos, freshwater-plankton, and freshwater-benthos

States <- c("mp", "mb", "fp", "fb")

# muhisse model
data("diatoms")
m_transition_matrix(
 model_fit = diatoms$muhisse,
 hidden_states = TRUE,
 states = States)

#>              mp1          mb1         fp1          fb1         mp2         mb2
#> mp1           NA 2.061154e-09 0.000125180 0.000000e+00 0.002832762          NA
#> mb1 7.971849e-04           NA 0.000000000 1.617498e-03          NA 0.002832762
#> fp1 2.061154e-09 0.000000e+00          NA 2.061154e-09          NA          NA
#> fb1 0.000000e+00 3.834735e-03 0.001416519           NA          NA          NA
#> mp2 6.483053e-02           NA          NA           NA          NA 0.002568749
#> mb2           NA 6.483053e-02          NA           NA 0.003050672          NA
#> fp2           NA           NA 0.064830535           NA 0.010208014 0.000000000
#> fb2           NA           NA          NA 6.483053e-02 0.000000000 0.002309284
#>              fp2         fb2
#> mp1           NA          NA
#> mb1           NA          NA
#> fp1 2.832762e-03          NA
#> fb1           NA 0.002832762
#> mp2 4.316343e-03 0.000000000
#> mb2 0.000000e+00 0.011032817
#> fp2           NA 0.031237128
#> fb2 2.061154e-09          NA

# CID8 model
# we have 8 hidden states, so the transition matrix is cumbersome
m_transition_matrix(
 model_fit = diatoms$cid8,
 hidden_states = TRUE,
 states = States)

#>              mp1         mb1          fp1         fb1          mp2         mb2
#> mp1           NA 0.001627666 0.0011089817 0.000000000 0.0061882626          NA
#> mb1 0.0008695831          NA 0.0000000000 0.005660987           NA 0.006188263
#> fp1 0.0021770116 0.000000000           NA 0.005455898           NA          NA
#> fb1 0.0000000000 0.002897852 0.0007733757          NA           NA          NA
#> mp2 0.0404592834          NA           NA          NA           NA 0.001627666
#> mb2           NA 0.040459283           NA          NA 0.0008695831          NA
#> fp2           NA          NA 0.0404592834          NA 0.0021770116 0.000000000
#> fb2           NA          NA           NA 0.040459283 0.0000000000 0.002897852
#> mp3 0.0404592834          NA           NA          NA 0.0061882626          NA
#> mb3           NA 0.040459283           NA          NA           NA 0.006188263
#> fp3           NA          NA 0.0404592834          NA           NA          NA
#> fb3           NA          NA           NA 0.040459283           NA          NA
#> mp4 0.0404592834          NA           NA          NA 0.0061882626          NA
#> mb4           NA 0.040459283           NA          NA           NA 0.006188263
#> fp4           NA          NA 0.0404592834          NA           NA          NA
#> fb4           NA          NA           NA 0.040459283           NA          NA
#> mp5 0.0404592834          NA           NA          NA 0.0061882626          NA
#> mb5           NA 0.040459283           NA          NA           NA 0.006188263
#> fp5           NA          NA 0.0404592834          NA           NA          NA
#> fb5           NA          NA           NA 0.040459283           NA          NA
#> mp6 0.0404592834          NA           NA          NA 0.0061882626          NA
#> mb6           NA 0.040459283           NA          NA           NA 0.006188263
#> fp6           NA          NA 0.0404592834          NA           NA          NA
#> fb6           NA          NA           NA 0.040459283           NA          NA
#> mp7 0.0404592834          NA           NA          NA 0.0061882626          NA
#> mb7           NA 0.040459283           NA          NA           NA 0.006188263
#> fp7           NA          NA 0.0404592834          NA           NA          NA
#> fb7           NA          NA           NA 0.040459283           NA          NA
#> mp8 0.0404592834          NA           NA          NA 0.0061882626          NA
#> mb8           NA 0.040459283           NA          NA           NA 0.006188263
#> fp8           NA          NA 0.0404592834          NA           NA          NA
#> fb8           NA          NA           NA 0.040459283           NA          NA
#>              fp2         fb2          mp3          mb3          fp3
#> mp1           NA          NA 4.399914e-08           NA           NA
#> mb1           NA          NA           NA 4.399914e-08           NA
#> fp1 0.0061882626          NA           NA           NA 4.399914e-08
#> fb1           NA 0.006188263           NA           NA           NA
#> mp2 0.0011089817 0.000000000 4.399914e-08           NA           NA
#> mb2 0.0000000000 0.005660987           NA 4.399914e-08           NA
#> fp2           NA 0.005455898           NA           NA 4.399914e-08
#> fb2 0.0007733757          NA           NA           NA           NA
#> mp3           NA          NA           NA 1.627666e-03 1.108982e-03
#> mb3           NA          NA 8.695831e-04           NA 0.000000e+00
#> fp3 0.0061882626          NA 2.177012e-03 0.000000e+00           NA
#> fb3           NA 0.006188263 0.000000e+00 2.897852e-03 7.733757e-04
#> mp4           NA          NA 4.399914e-08           NA           NA
#> mb4           NA          NA           NA 4.399914e-08           NA
#> fp4 0.0061882626          NA           NA           NA 4.399914e-08
#> fb4           NA 0.006188263           NA           NA           NA
#> mp5           NA          NA 4.399914e-08           NA           NA
#> mb5           NA          NA           NA 4.399914e-08           NA
#> fp5 0.0061882626          NA           NA           NA 4.399914e-08
#> fb5           NA 0.006188263           NA           NA           NA
#> mp6           NA          NA 4.399914e-08           NA           NA
#> mb6           NA          NA           NA 4.399914e-08           NA
#> fp6 0.0061882626          NA           NA           NA 4.399914e-08
#> fb6           NA 0.006188263           NA           NA           NA
#> mp7           NA          NA 4.399914e-08           NA           NA
#> mb7           NA          NA           NA 4.399914e-08           NA
#> fp7 0.0061882626          NA           NA           NA 4.399914e-08
#> fb7           NA 0.006188263           NA           NA           NA
#> mp8           NA          NA 4.399914e-08           NA           NA
#> mb8           NA          NA           NA 4.399914e-08           NA
#> fp8 0.0061882626          NA           NA           NA 4.399914e-08
#> fb8           NA 0.006188263           NA           NA           NA
#>              fb3          mp4         mb4          fp4         fb4          mp5
#> mp1           NA 0.0071726133          NA           NA          NA 0.0001173700
#> mb1           NA           NA 0.007172613           NA          NA           NA
#> fp1           NA           NA          NA 0.0071726133          NA           NA
#> fb1 4.399914e-08           NA          NA           NA 0.007172613           NA
#> mp2           NA 0.0071726133          NA           NA          NA 0.0001173700
#> mb2           NA           NA 0.007172613           NA          NA           NA
#> fp2           NA           NA          NA 0.0071726133          NA           NA
#> fb2 4.399914e-08           NA          NA           NA 0.007172613           NA
#> mp3 0.000000e+00 0.0071726133          NA           NA          NA 0.0001173700
#> mb3 5.660987e-03           NA 0.007172613           NA          NA           NA
#> fp3 5.455898e-03           NA          NA 0.0071726133          NA           NA
#> fb3           NA           NA          NA           NA 0.007172613           NA
#> mp4           NA           NA 0.001627666 0.0011089817 0.000000000 0.0001173700
#> mb4           NA 0.0008695831          NA 0.0000000000 0.005660987           NA
#> fp4           NA 0.0021770116 0.000000000           NA 0.005455898           NA
#> fb4 4.399914e-08 0.0000000000 0.002897852 0.0007733757          NA           NA
#> mp5           NA 0.0071726133          NA           NA          NA           NA
#> mb5           NA           NA 0.007172613           NA          NA 0.0008695831
#> fp5           NA           NA          NA 0.0071726133          NA 0.0021770116
#> fb5 4.399914e-08           NA          NA           NA 0.007172613 0.0000000000
#> mp6           NA 0.0071726133          NA           NA          NA 0.0001173700
#> mb6           NA           NA 0.007172613           NA          NA           NA
#> fp6           NA           NA          NA 0.0071726133          NA           NA
#> fb6 4.399914e-08           NA          NA           NA 0.007172613           NA
#> mp7           NA 0.0071726133          NA           NA          NA 0.0001173700
#> mb7           NA           NA 0.007172613           NA          NA           NA
#> fp7           NA           NA          NA 0.0071726133          NA           NA
#> fb7 4.399914e-08           NA          NA           NA 0.007172613           NA
#> mp8           NA 0.0071726133          NA           NA          NA 0.0001173700
#> mb8           NA           NA 0.007172613           NA          NA           NA
#> fp8           NA           NA          NA 0.0071726133          NA           NA
#> fb8 4.399914e-08           NA          NA           NA 0.007172613           NA
#>             mb5          fp5         fb5          mp6         mb6          fp6
#> mp1          NA           NA          NA 0.0014716704          NA           NA
#> mb1 0.000117370           NA          NA           NA 0.001471670           NA
#> fp1          NA 0.0001173700          NA           NA          NA 0.0014716704
#> fb1          NA           NA 0.000117370           NA          NA           NA
#> mp2          NA           NA          NA 0.0014716704          NA           NA
#> mb2 0.000117370           NA          NA           NA 0.001471670           NA
#> fp2          NA 0.0001173700          NA           NA          NA 0.0014716704
#> fb2          NA           NA 0.000117370           NA          NA           NA
#> mp3          NA           NA          NA 0.0014716704          NA           NA
#> mb3 0.000117370           NA          NA           NA 0.001471670           NA
#> fp3          NA 0.0001173700          NA           NA          NA 0.0014716704
#> fb3          NA           NA 0.000117370           NA          NA           NA
#> mp4          NA           NA          NA 0.0014716704          NA           NA
#> mb4 0.000117370           NA          NA           NA 0.001471670           NA
#> fp4          NA 0.0001173700          NA           NA          NA 0.0014716704
#> fb4          NA           NA 0.000117370           NA          NA           NA
#> mp5 0.001627666 0.0011089817 0.000000000 0.0014716704          NA           NA
#> mb5          NA 0.0000000000 0.005660987           NA 0.001471670           NA
#> fp5 0.000000000           NA 0.005455898           NA          NA 0.0014716704
#> fb5 0.002897852 0.0007733757          NA           NA          NA           NA
#> mp6          NA           NA          NA           NA 0.001627666 0.0011089817
#> mb6 0.000117370           NA          NA 0.0008695831          NA 0.0000000000
#> fp6          NA 0.0001173700          NA 0.0021770116 0.000000000           NA
#> fb6          NA           NA 0.000117370 0.0000000000 0.002897852 0.0007733757
#> mp7          NA           NA          NA 0.0014716704          NA           NA
#> mb7 0.000117370           NA          NA           NA 0.001471670           NA
#> fp7          NA 0.0001173700          NA           NA          NA 0.0014716704
#> fb7          NA           NA 0.000117370           NA          NA           NA
#> mp8          NA           NA          NA 0.0014716704          NA           NA
#> mb8 0.000117370           NA          NA           NA 0.001471670           NA
#> fp8          NA 0.0001173700          NA           NA          NA 0.0014716704
#> fb8          NA           NA 0.000117370           NA          NA           NA
#>             fb6          mp7         mb7          fp7         fb7          mp8
#> mp1          NA 0.0002479810          NA           NA          NA 0.0070178810
#> mb1          NA           NA 0.000247981           NA          NA           NA
#> fp1          NA           NA          NA 0.0002479810          NA           NA
#> fb1 0.001471670           NA          NA           NA 0.000247981           NA
#> mp2          NA 0.0002479810          NA           NA          NA 0.0070178810
#> mb2          NA           NA 0.000247981           NA          NA           NA
#> fp2          NA           NA          NA 0.0002479810          NA           NA
#> fb2 0.001471670           NA          NA           NA 0.000247981           NA
#> mp3          NA 0.0002479810          NA           NA          NA 0.0070178810
#> mb3          NA           NA 0.000247981           NA          NA           NA
#> fp3          NA           NA          NA 0.0002479810          NA           NA
#> fb3 0.001471670           NA          NA           NA 0.000247981           NA
#> mp4          NA 0.0002479810          NA           NA          NA 0.0070178810
#> mb4          NA           NA 0.000247981           NA          NA           NA
#> fp4          NA           NA          NA 0.0002479810          NA           NA
#> fb4 0.001471670           NA          NA           NA 0.000247981           NA
#> mp5          NA 0.0002479810          NA           NA          NA 0.0070178810
#> mb5          NA           NA 0.000247981           NA          NA           NA
#> fp5          NA           NA          NA 0.0002479810          NA           NA
#> fb5 0.001471670           NA          NA           NA 0.000247981           NA
#> mp6 0.000000000 0.0002479810          NA           NA          NA 0.0070178810
#> mb6 0.005660987           NA 0.000247981           NA          NA           NA
#> fp6 0.005455898           NA          NA 0.0002479810          NA           NA
#> fb6          NA           NA          NA           NA 0.000247981           NA
#> mp7          NA           NA 0.001627666 0.0011089817 0.000000000 0.0070178810
#> mb7          NA 0.0008695831          NA 0.0000000000 0.005660987           NA
#> fp7          NA 0.0021770116 0.000000000           NA 0.005455898           NA
#> fb7 0.001471670 0.0000000000 0.002897852 0.0007733757          NA           NA
#> mp8          NA 0.0002479810          NA           NA          NA           NA
#> mb8          NA           NA 0.000247981           NA          NA 0.0008695831
#> fp8          NA           NA          NA 0.0002479810          NA 0.0021770116
#> fb8 0.001471670           NA          NA           NA 0.000247981 0.0000000000
#>             mb8          fp8         fb8
#> mp1          NA           NA          NA
#> mb1 0.007017881           NA          NA
#> fp1          NA 0.0070178810          NA
#> fb1          NA           NA 0.007017881
#> mp2          NA           NA          NA
#> mb2 0.007017881           NA          NA
#> fp2          NA 0.0070178810          NA
#> fb2          NA           NA 0.007017881
#> mp3          NA           NA          NA
#> mb3 0.007017881           NA          NA
#> fp3          NA 0.0070178810          NA
#> fb3          NA           NA 0.007017881
#> mp4          NA           NA          NA
#> mb4 0.007017881           NA          NA
#> fp4          NA 0.0070178810          NA
#> fb4          NA           NA 0.007017881
#> mp5          NA           NA          NA
#> mb5 0.007017881           NA          NA
#> fp5          NA 0.0070178810          NA
#> fb5          NA           NA 0.007017881
#> mp6          NA           NA          NA
#> mb6 0.007017881           NA          NA
#> fp6          NA 0.0070178810          NA
#> fb6          NA           NA 0.007017881
#> mp7          NA           NA          NA
#> mb7 0.007017881           NA          NA
#> fp7          NA 0.0070178810          NA
#> fb7          NA           NA 0.007017881
#> mp8 0.001627666 0.0011089817 0.000000000
#> mb8          NA 0.0000000000 0.005660987
#> fp8 0.000000000           NA 0.005455898
#> fb8 0.002897852 0.0007733757          NA

# musse model, no hidden states
m_transition_matrix(
 model_fit = diatoms$musse,
 hidden_states = FALSE,
 states = c("mp", "mb", "fp", "fb"))

#>             mp          mb           fp           fb
#> mp          NA 0.001321135 0.0012224284 0.0000000000
#> mb 0.001125972          NA 0.0000000000 0.0001720078
#> fp 0.002168149 0.000000000           NA 0.0086348164
#> fb 0.000000000 0.007665693 0.0006466656           NA